Add sorting and remove reducer
