+++ packages/common/src/operations.graphql
@@ -195,8 +195,11 @@ query invoice($id: String!) {
         ... on InvoicePricingNode {
           __typename
           totalAfterTax
-          # subtotal
-          # taxPercentage
+          totalBeforeTax
+          stockTotalBeforeTax
+          stockTotalAfterTax
+          serviceTotalAfterTax
+          serviceTotalBeforeTax
         }
       }
       status
@@ -650,8 +653,11 @@ query invoices(
           ... on InvoicePricingNode {
             __typename
             totalAfterTax
-            # subtotal
-            # taxPercentage
+            totalBeforeTax
+            stockTotalBeforeTax
+            stockTotalAfterTax
+            serviceTotalAfterTax
+            serviceTotalBeforeTax
           }
         }
       }
