--- packages/invoices/src/InboundShipment/DetailView/reducer/index.ts
@@ -1,1 +1,0 @@
-export * from './reducer';
