--- packages/invoices/src/InboundShipment/DetailView/reducer/reducer.test.ts
@@ -1,3 +1,0 @@
-describe('Inbound shipment DetailView Reducer', () => {
-  it('is a placeholder', () => {});
-});
