+++ packages/invoices/src/utils.ts
@@ -1,8 +1,13 @@
 import {
+  getUnitQuantity,
+  getSumOfKeyReducer,
   LocaleKey,
   InvoiceNodeStatus,
   useTranslation,
   InvoiceNodeType,
+  groupBy,
+  ifTheSameElseDefault,
+  arrayToRecord,
 } from '@openmsupply-client/common';
 import {
   OutboundShipment,
@@ -12,6 +17,7 @@ import {
   InboundShipmentItem,
   InboundShipmentRow,
   Invoice,
+  InvoiceLine,
 } from './types';
 
 export const placeholderInbound: InboundShipment = {
@@ -210,4 +216,34 @@ export const flattenInboundItems = (
   summaryItems: InboundShipmentItem[]
 ): InboundShipmentRow[] => {
   return summaryItems.map(({ batches }) => Object.values(batches)).flat();
+};
+
+export const createSummaryItem = (
+  itemId: string,
+  batches: InboundShipmentRow[] = []
+): InboundShipmentItem => {
+  const item: InboundShipmentItem = {
+    id: itemId,
+    itemId: itemId,
+    itemName: ifTheSameElseDefault(batches, 'itemName', ''),
+    itemCode: ifTheSameElseDefault(batches, 'itemCode', ''),
+    batches: arrayToRecord(batches),
+    unitQuantity: batches.reduce(getUnitQuantity, 0),
+    numberOfPacks: batches.reduce(getSumOfKeyReducer('numberOfPacks'), 0),
+    locationName: ifTheSameElseDefault(batches, 'locationName', undefined),
+    batch: ifTheSameElseDefault(batches, 'batch', '[multiple]'),
+    sellPrice: ifTheSameElseDefault(batches, 'sellPricePerPack', undefined),
+    packSize: ifTheSameElseDefault(batches, 'packSize', undefined),
+  };
+
+  return item;
+};
+
+export const inboundLinesToSummaryItems = (
+  lines: InvoiceLine[]
+): OutboundShipmentSummaryItem[] => {
+  const grouped = groupBy(lines, 'itemId');
+  return Object.keys(grouped).map(itemId =>
+    createSummaryItem(itemId, grouped[itemId])
+  );
 };
